+++ com.unity.shadergraph/Editor/Data/Graphs/AbstractShaderProperty.cs
@@ -64,6 +64,19 @@ namespace UnityEditor.ShaderGraph
             set { m_GeneratePropertyBlock = value; }
         }
 
+        private ConcretePrecision m_ConcretePrecision = ConcretePrecision.Float;
+
+        public ConcretePrecision concretePrecision => m_ConcretePrecision;
+
+        [SerializeField]
+        private Precision m_Precision = Precision.Inherit;
+
+        public Precision precision
+        {
+            get => m_Precision;
+            set => m_Precision = value;
+        }
+
         public abstract PropertyType propertyType { get; }
         
         public abstract Vector4 defaultValue { get; }
@@ -76,6 +89,11 @@ namespace UnityEditor.ShaderGraph
         public virtual string GetPropertyAsArgumentString()
         {
             return GetPropertyDeclarationString(string.Empty);
+        }
+
+        public void SetConcretePrecision(ConcretePrecision inheritedPrecision)
+        {
+            m_ConcretePrecision = (precision == Precision.Inherit) ? inheritedPrecision : precision.ToConcrete();
         }
 
         public abstract PreviewProperty GetPreviewMaterialProperty();