+++ com.unity.shadergraph/Editor/Data/Nodes/Input/Basic/IntegerNode.cs
@@ -61,7 +61,7 @@ namespace UnityEditor.ShaderGraph
             if (generationMode.IsPreview())
                 return;
 
-            sb.AppendLine(precision + " " + GetVariableNameForNode() + " = " + m_Value + ";");
+            sb.AppendLine(string.Format("$precision {0} = {1};", GetVariableNameForNode(), m_Value));
         }
 
         public override string GetVariableNameForSlot(int slotId)