+++ com.unity.shadergraph/Editor/Drawing/Views/MaterialNodeView.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Reflection;
 using UnityEngine;
 using UnityEditor.Graphing;
+using UnityEditor.Graphing.Util;
 using UnityEditor.ShaderGraph.Drawing.Controls;
 using UnityEngine.Rendering;
 
@@ -11,6 +12,7 @@ using UnityEditor.Experimental.GraphView;
 using UnityEditor.Rendering;
 using UnityEditor.ShaderGraph.Drawing.Colors;
 using UnityEngine.UIElements;
+using UnityEditor.UIElements;
 using Node = UnityEditor.Experimental.GraphView.Node;
 
 namespace UnityEditor.ShaderGraph.Drawing
@@ -151,6 +153,10 @@ namespace UnityEditor.ShaderGraph.Drawing
                 RegisterCallback<MouseDownEvent>(OnSubGraphDoubleClick);
             }
 
+            m_PortInputContainer.SendToBack();
+
+            m_TitleContainer = this.Q("title");
+
             var masterNode = node as IMasterNode;
             if (masterNode != null)
             {
@@ -162,28 +168,29 @@ namespace UnityEditor.ShaderGraph.Drawing
                 }
             }
 
-            m_PortInputContainer.SendToBack();
+            m_NodeSettingsView = new NodeSettingsView();
+            m_NodeSettingsView.visible = false;
+            Add(m_NodeSettingsView);
 
-            m_TitleContainer = this.Q("title");
+            m_SettingsButton = new VisualElement {name = "settings-button"};
+            m_SettingsButton.Add(new VisualElement { name = "icon" });
 
-            var settings = node as IHasSettings;
-            if (settings != null)
+            m_Settings = new VisualElement();
+            AddDefaultSettings();
+
+            // Add Node type specific settings
+            var nodeTypeSettings = node as IHasSettings;
+            if (nodeTypeSettings != null)
+                m_Settings.Add(nodeTypeSettings.CreateSettingsElement());
+            
+            // Add manipulators
+            m_SettingsButton.AddManipulator(new Clickable(() =>
+                {
+                    UpdateSettingsExpandedState();
+                }));
+
+            if(m_Settings.childCount > 0)
             {
-                m_NodeSettingsView = new NodeSettingsView();
-                m_NodeSettingsView.visible = false;
-
-                Add(m_NodeSettingsView);
-
-                m_SettingsButton = new VisualElement {name = "settings-button"};
-                m_SettingsButton.Add(new VisualElement { name = "icon" });
-
-                m_Settings = settings.CreateSettingsElement();
-
-                m_SettingsButton.AddManipulator(new Clickable(() =>
-                    {
-                        UpdateSettingsExpandedState();
-                    }));
-
                 m_ButtonContainer = new VisualElement { name = "button-container" };
                 m_ButtonContainer.style.flexDirection = FlexDirection.Row;
                 m_ButtonContainer.Add(m_SettingsButton);
@@ -352,17 +359,57 @@ namespace UnityEditor.ShaderGraph.Drawing
             return node.owner.GetShader(node, mode, node.name).shader;
         }
 
+        void AddDefaultSettings()
+        {
+            PropertySheet ps = new PropertySheet();
+            bool hasDefaultSettings = false;
+
+            if(node.canSetPrecision)
+            {
+                hasDefaultSettings = true;
+                ps.Add(new PropertyRow(new Label("Precision")), (row) =>
+                {
+                    row.Add(new EnumField(node.precision), (field) =>
+                    {
+                        field.RegisterValueChangedCallback(evt =>
+                        {
+                            if (evt.newValue.Equals(node.precision))
+                                return;
+                            
+                            var editorView = GetFirstAncestorOfType<GraphEditorView>();
+                            var nodeList = m_GraphView.Query<MaterialNodeView>().ToList();
+
+                            editorView.colorManager.SetNodesDirty(nodeList);
+                            node.owner.owner.RegisterCompleteObjectUndo("Change precision");
+                            node.precision = (Precision)evt.newValue;
+                            node.owner.ValidateGraph();
+                            editorView.colorManager.UpdateNodeViews(nodeList);
+                            node.Dirty(ModificationScope.Graph);
+                        });
+                    });
+                });
+            }
+
+            if(hasDefaultSettings)
+                m_Settings.Add(ps);
+        }
+
         void RecreateSettings()
         {
-            var settings = node as IHasSettings;
-            if (settings != null)
-            {
-                m_Settings.RemoveFromHierarchy();
+            m_Settings.RemoveFromHierarchy();
+            m_Settings = new PropertySheet();
 
-                m_Settings = settings.CreateSettingsElement();
-                m_NodeSettingsView.Add(m_Settings);
-            }
+            // Add default settings
+            AddDefaultSettings();
+
+            // Add Node type specific settings
+            var nodeTypeSettings = node as IHasSettings;
+            if (nodeTypeSettings != null)
+                m_Settings.Add(nodeTypeSettings.CreateSettingsElement());
+
+            m_NodeSettingsView.Add(m_Settings);
         }
+
 
         void UpdateSettingsExpandedState()
         {