+++ com.unity.shadergraph/CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [6.8.0] - 2019-XX-XX
 ### Added
-- Color Mode improves node visibility by coloring the title bar by Category, Precision, or custom colors
+- Color Mode improves node visibility by coloring the title bar by Category, Precision, or custom colors.
+- You can now set the precision of a Shader Graph and individual nodes.
 
 ## [6.7.0] - 2019-05-21
 ### Added